oracle-monitor/cli/requirements.txt
click==8.1.7
supabase==2.3.0
deepdiff==6.7.1
python-dotenv==1.0.1
rich==13.7.0
